add estimation w/ act now params + organize notebook
