+++ notebooks/themes/custom_colorscales.yaml
@@ -1,0 +1,9 @@
+
+# Add custom colorscales here!
+my_colors: ['#DA4657', '#054257', '#49B5BD', '#B1F200', '#FEB916']
+
+# Source: https://learnui.design/tools/data-color-picker.html
+learn_ui_design: ['#003f5c', '#58508d', '#bc5090', '#ff6361', '#ffa600'] 
+
+# R_t
+Rt: ['#da4657', '#d1d1d1', '#d1d1d1']
